refactor: tabStore 객체 위치 수정에 따라 파일 경로 변경
